+++ R/diffnet-class.R
@@ -1,3 +1,300 @@
+# Creates n column names
+make_col_names <- function(n, is.dynamic, prefix="v.") {
+  expression <- paste0("%s%s%0", nchar(n), "d")
+  if (is.dynamic) sprintf(expression,prefix,"dyn.", seq_len(n))
+  else sprintf(expression,prefix,"static.", seq_len(n))
+}
+
+# Checks attributes to be added to a diffnet object
+check_as_diffnet_attrs <- function(attrs, meta, is.dynamic, id.and.per.vars=NULL,
+                                   attr.class="vertex") {
+  # Getting meta
+  n    <- meta$n
+  nper <- meta$nper
+
+  # In case of merging
+  iddf  <- data.frame(id  = meta$ids)
+  perdf <- data.frame(per = meta$pers)
+
+  # Vertex Attrubutes  ---------------------------------------------------------
+  if (attr.class == "vertex") {
+    if (is.dynamic) {
+      # If null
+      if (!length(attrs)) {
+        attrs <-
+          lapply(meta$pers, function(x) {
+            as.data.frame(
+              matrix(ncol=0, nrow=n, dimnames = list(meta$ids, NULL))
+            )
+          })
+
+        # Labeling
+        names(attrs) <- meta$pers
+        return(attrs)
+      }
+
+
+      if (inherits(attrs, "list")) {
+        # Must have the same number of elements as time periods
+        if (length(attrs) != nper)
+          stop("The list -vertex.dyn.attrs-'s length must be equal to the number of slices.")
+
+        # Checking that all are data.frames/matrices or vectors
+        isdf  <- sapply(attrs, inherits, what="data.frame")
+        ismat <- sapply(attrs, inherits, what="matrix")
+        isvec <- sapply(attrs, is.vector)
+
+        test <- which(!isdf & !ismat & !isvec)
+        if (length(test))
+          stop("Some in the list -vertex.dyn.attrs- are not supported:\n\t:",
+               paste0(test, collapse=", "), ".")
+
+        # Further, all should be of the same class
+        if ( (any(isdf) + any(ismat) + any(isvec))>1 )
+          stop("All elements in the list -vertex.dyn.attrs- should be of the same class.")
+
+        if (all(isdf) | all(ismat)) {
+          # Dynamic: List of matrices/data.frames --------------------------------
+          # Checking dims
+          test <- which(sapply(attrs, nrow) != n)
+          if (length(test)) {
+            stop("Some matrices/data frames in -vertex.dyn.attrs- don't have n (",
+                 n, ") rows:\n\t", paste0(test, collapse = ", "), ".")
+          }
+
+          # Checking number of columns
+          test <- which(attrs[[1]] == sapply(attrs, ncol))
+          if (length(test))
+            stop("All elements in -vertex.dyn.attrs- must have the same number of columns.")
+
+          # Checking columnnames
+          test <- colnames(attrs[[1]])
+          test <- if (length(test)) which(sapply(attrs, function(x) all(colnames(x) != test)))
+          else which(sapply(attrs, function(x) length(colnames(x))))
+
+          if (length(test))
+            stop("Not all the matrices/data frames in -vertex.dyn.attrs- have the same name:\n\t",
+                 paste0(test, collapse = ", "))
+
+          # Making up names
+          cnames <- colnames(attrs[[1]])
+          if (!length(cnames))
+            cnames <- make_col_names(ncol(attrs[[1]]), TRUE)
+
+          # Returning as data.frame
+          attrs <- lapply(attrs, function(x) {
+            x <- as.data.frame(x)
+
+            # Checking if must be ordered
+            if (length(id.and.per.vars)) {
+              x <- merge(iddf, x, by.x="id", by.y=id.and.per.vars[1],
+                         sort=FALSE, all.x=TRUE, all.y=FALSE)
+
+              # Sorting names back
+              colnames(x)[1] <- id.and.per.vars[1]
+              x <- x[,cnames]
+              x
+            }
+
+            # Naming
+            dimnames(x) <- list(meta$ids, cnames)
+            x
+            })
+
+          # Adding names
+          names(attrs) <- meta$pers
+
+          return(attrs)
+
+        } else { # Dynamic: List of vectors --------------------------------------
+          # Checking dim
+          test <- which(sapply(attrs, length) != n*nper)
+          if (length(test))
+            stop("Some vectors in -vertex.dyn.attrs- have a different lengths than expected:\n\t",
+                 paste0(test, collapse=", "), ".")
+
+          cnames <- make_col_names(1, TRUE)
+
+          # Returning as data.frame
+          attrs <- lapply(attrs, function(x) {
+            x <- as.data.frame(x)
+
+            dimnames(x) <- list(meta$ids, cnames)
+            x
+            })
+
+          # Adding names
+          names(attrs) <- meta$pers
+
+          return(attrs)
+        }
+
+      } else if (inherits(attrs, "data.frame") | inherits(attrs, "matrix")) {
+        # Dynamic: Matrix or data.frame n*nper -----------------------------------
+        if (nrow(attrs) != n*nper)
+          stop("The matrix/data.frame -vertex.dyn.attrs- has incorrect number of rows.")
+
+        # Checking colnames
+        cnames <- colnames(attrs)
+        if (!length(cnames))
+          cnames <- make_col_names(ncol(attrs), TRUE)
+
+        # Coercing into a list of data.frames
+        attrs <- lapply(meta$pers, function(x) {
+
+          # Subseting
+          x <- if (!length(id.and.per.vars[2])) attrs[((x-1)*n + 1):(x*n),,drop=FALSE]
+          else attrs[attrs[,id.and.per.vars[2]] == x,,drop=FALSE]
+
+          x <- as.data.frame(x)
+
+          # Checking if must be ordered
+          if (length(id.and.per.vars)) {
+            x <- merge(iddf, x, by.x="id", by.y=id.and.per.vars[1],
+                       sort=FALSE, all.x=TRUE, all.y=FALSE)
+
+            # Sorting names (back)
+            colnames(x)[1] <- id.and.per.vars[1]
+            x <- x[,cnames]
+            x
+          }
+
+          dimnames(x) <- list(meta$ids, cnames)
+          x}
+        )
+
+        # Adding names
+        names(attrs) <- meta$pers
+
+        # Returning
+        return(attrs)
+
+      } else if (is.vector(attrs)) {
+        # Dynamic: vector of size n*nper -----------------------------------------
+        if (length(attrs) != n*nper)
+          stop("The vector -vertex.dyn.attrs- has incorrect length.")
+
+        # Colname
+        cnames <- make_col_names(1, TRUE)
+
+        # Coercing into a list of data.frames
+        attrs <- lapply(meta$pers, function(x) {
+          x <- as.data.frame(attrs[((x-1)*n + 1):(x*n)])
+          dimnames(x) <- list(meta$ids, cnames)
+          x}
+        )
+
+        # Adding names
+        names(attrs) <- meta$pers
+
+        # Returning
+        return(attrs)
+
+      } else
+        stop("The class of -vertex.dyn.attrs-, \'", class(attrs),
+             "\', is not supported.")
+
+    } else { # Static attributes
+      # If null
+      if (!length(attrs)) {
+        return(as.data.frame(
+          matrix(ncol=0, nrow=n, dimnames = list(meta$ids, NULL))
+        ))
+      }
+
+      if (inherits(attrs, "data.frame") | inherits(attrs, "matrix")) {
+        # Static: Data frame or matrix -------------------------------------------
+        if (nrow(attrs) != n)
+          stop("The matrix/data.frame -vertex.static.attrs- has incorrect number of rows.")
+
+        # If we need to sort using id
+        attrs <- as.data.frame(attrs)
+        cnames <- colnames(attrs)
+        if (length(id.and.per.vars)) {
+          attrs <- merge(iddf, attrs, by.x="id", by.y=id.and.per.vars[1],
+                     sort=FALSE, all.x=TRUE, all.y=FALSE)
+
+          # Sorting names back
+          colnames(attrs)[1] <- id.and.per.vars[1]
+          attrs <- attrs[,cnames]
+        }
+
+        # Checking colnames
+        if (!length(cnames))
+          cnames <- make_col_names(ncol(attrs), FALSE)
+
+        # Returning
+        dimnames(attrs) <- list(meta$ids, cnames)
+        return(attrs)
+
+      } else if (is.vector(attrs)) {
+        # Static: Vector -------------------------------------------------------
+        if (length(attrs) != n)
+          stop("The vector -vertex.static.attrs- has incorrect length.")
+
+        # Returning
+        attrs <- as.data.frame(attrs)
+        dimnames(attrs) <- list(meta$ids, make_col_names(1, FALSE))
+
+        return(attrs)
+
+      } else
+        stop("The class of -vertex.static.attrs-, \'", class(attrs),
+             "\', is not supported.")
+
+    }
+  } else if (attr.class == "graph") {
+    # if Null
+    if (!length(attrs)) {
+      return(as.data.frame(
+          matrix(ncol=0, nrow=nper, dimnames = list(meta$pers, NULL))
+        ))
+    }
+
+    # Graph attributes ---------------------------------------------------------
+    if (inherits(attrs, "data.frame") | inherits(attrs, "matrix")) {
+      # Checking number of rows
+      if (nrow(attrs) != nper)
+        stop("The matrix/data.frame -graph.attrs- has incorrect number of rows.")
+
+      # Checking the order of the data
+      attrs <- as.data.frame(attrs)
+      cnames <- colnames(attrs)
+      if (length(id.and.per.vars[2])) {
+        attrs <- merge(perdf, attrs, by.x="per", by.y=id.and.per.vars[2],
+                       all.x=TRUE, all.y=FALSE)
+        # Sorting names back
+        colnames(attrs)[1] <- id.and.per.vars[1]
+        attrs <- attrs[,cnames]
+      }
+
+      # Checking colnames
+      if (!length(cnames))
+        cnames <- make_col_names(ncol(attrs), TRUE, prefix = "graph")
+
+      # Returning
+      dimnames(attrs) <- list(meta$pers, cnames)
+
+      return(attrs)
+
+    } else if (is.vector(attrs)) {
+
+      # Checking the length
+      if (length(attrs) != nper)
+        stop("The vector -graph.attrs- has incorrect length.")
+
+      # Returning
+      attrs <- as.data.frame(attrs)
+      dimnames(attrs) <- list(meta$pers, make_col_names(1, TRUE, prefix = "graph"))
+
+      return(attrs)
+
+    } else
+      stop("The class of -graph.attrs-, \'", class(attrs),
+           "\', is not supported.")
+  }
+}
+
 #' Creates a \code{diffnet} class object
 #'
 #' \code{diffnet} objects contain difussion of innovation networks. With adjacency
@@ -27,10 +324,11 @@
 #' @param main Character. A title template to be passed to sprintf.
 #' @param i Indices specifying elements to replace. See \code{\link[base:Extract]{Extract}}.
 #' @param value In the case of \code{diffnet.toa}, replacement, otherwise see below.
-#' @param vertex.dyn.attrs List of length \eqn{T}. Contains either matrices or data
-#' frames with vertex attributes.
-#' @param vertex.static.attrs Either a data frame or a matrix with \eqn{n} rows.
-#' @param graph.attrs Numeric matrix with a single row.
+#' @param vertex.dyn.attrs Vertices dynamic attributes (see details).
+#' @param vertex.static.attrs Vertices static attributes (see details).
+#' @param graph.attrs Graph dynamic attributes (not supported yet).
+#' @param id.and.per.vars A character vector of length 2. Optionally specified to check the
+#' order of the rows in the attribute data.
 #' @param slices Either an integer or character vector. While integer vectors are used as
 #' indexes, character vectors are used jointly with the time period labels.
 #' @param element Character vector/scalar. Indicates what to retrieve/alter.
@@ -42,7 +340,38 @@
 #' @param valued Logical scalar. When FALSE non-zero values in the adjmat are set to one.
 #' @export
 #' @seealso Default options are listed at \code{\link{netdiffuseR-options}}
-#' @details Plotting is done via the function \code{\link[sna:gplot]{gplot}},
+#' @details
+#'
+#' \code{diffnet} objects hold both, static and dynamic vertex attributes. When
+#' creating \code{diffnet} objects, these can be specified using the arguments
+#' \code{vertex.static.attrs} and \code{vertex.dyn.attrs}; depending on whether
+#' the attributes to specify are static or dynamic, \pkg{netdiffuseR} currently
+#' supports the following objects:
+#'
+#' \tabular{llr}{
+#' \strong{Class}    \tab \strong{Dimension}               \tab \strong{Check sorting}\cr
+#' \emph{Static attributes} \cr
+#' \code{matrix}     \tab with \eqn{n} rows                \tab \code{id} \cr
+#' \code{data.frame} \tab with \eqn{n} rows                \tab \code{id} \cr
+#' \code{vector}     \tab of length \eqn{n}                \tab - \cr\cr
+#' \emph{Dynamic attributes} \cr
+#' \code{matrix}     \tab with \eqn{n\times T}{n * T} rows \tab \code{id}, \code{per} \cr
+#' \code{data.frame} \tab with \eqn{n\times T}{n * T} rows \tab \code{id}, \code{per}\cr
+#' \code{vector}     \tab of length \eqn{n\times T}{n*T}   \tab - \cr
+#' \code{list}       \tab of length \eqn{T} with matrices or data.frames of \eqn{n} rows\tab \code{id}, \code{per}\cr
+#' }
+#'
+#' The last column, \strong{Check sorting}, lists the variables that
+#' the user should specify if he wants the function to check the order of the rows
+#' of the attributes (notice that this is not possible for the case of vectors).
+#' By providing the name of the vertex id variable, \code{id}, and the time period
+#' id variable, \code{per}, the function makes sure that the attribute data is
+#' presented in the right order. See the example below. If the user does not
+#' provide the names of the vertex id and time period variables then the function
+#' does not check the way the rows are sorted, further it assumes that the data
+#' is in the correct order.
+#'
+#' Plotting is done via the function \code{\link[sna:gplot]{gplot}},
 #' and its layout via \code{\link[sna:gplot.layout]{gplot.layout}}, both from
 #' the (\pkg{sna}) package.
 #'
@@ -102,6 +431,7 @@
 #'
 #' @family diffnet methods
 #' @aliases diffnet diffnet-class
+#' @seealso A nice wrapper of this function is \code{\link{survey_to_diffnet}}.
 #' @examples
 #'
 #' # Creating a random graph
@@ -132,6 +462,36 @@
 #' # Now as a data.frame (all of them)
 #' diffnet.attrs(diffnet, as.df = TRUE)
 #'
+#' # Unsorted data -------------------------------------------------------------
+#' # Loading example data
+#' data(fakesurveyDyn)
+#'
+#' # Creating a diffnet object
+#' fs_diffnet <- survey_to_diffnet(
+#'    fakesurveyDyn, "id", c("net1", "net2", "net3"), "toa", "group",
+#'    timevar = "time", keep.isolates=TRUE, warn.coercion=FALSE)
+#'
+#' # Now, we extract the graph data and create a diffnet object from scratch
+#' graph <- fs_diffnet$graph
+#' attrs <- diffnet.attrs(fs_diffnet, as.df=TRUE)
+#' toa   <- diffnet.toa(fs_diffnet)
+#'
+#' # Lets apply a different sorting to the data to see if it works
+#' n <- nrow(attrs)
+#' attrs <- attrs[order(runif(n)),]
+#'
+#' # Now, recreating the old diffnet object (notice -id.and.per.vars- arg)
+#' fs_diffnet_new <- as_diffnet(graph, toa=toa, vertex.dyn.attrs=attrs,
+#'    id.and.per.vars = c("id", "per"))
+#'
+#' # Now, retrieving attributes. The 'new one' will have more (repeated)
+#' attrs_new <- diffnet.attrs(fs_diffnet_new, as.df=TRUE)
+#' attrs_old <- diffnet.attrs(fs_diffnet, as.df=TRUE)
+#'
+#' # Comparing elements!
+#' tocompare <- intersect(colnames(attrs_new), colnames(attrs_old))
+#' all(attrs_new[,tocompare] == attrs_old[,tocompare], na.rm = TRUE) # TRUE!
+#'
 #' @return
 #' A list of class \code{diffnet} with the following elements:
 #' \item{graph}{A list of length \eqn{T}. Containing sparse square matrices of size \eqn{n}
@@ -143,7 +503,7 @@
 #' attributes.}
 #' \item{vertex.dyn.attrs}{A list of length \eqn{T} with data frames containing vertex attributes
 #' throught time (dynamic).}
-#' \item{graph.attrs}{If not NULL, a numeric matrix with 1 row containing graph attributes.}
+#' \item{graph.attrs}{A data frame with \eqn{T} rows.}
 #' \item{meta}{A list of length 9 with the following elements:
 #' \itemize{
 #'  \item \code{type}: Character scalar equal to \code{"dynamic"}.
@@ -160,11 +520,12 @@
 #' @author Vega Yon
 as_diffnet <- function(graph, toa, t0=min(toa, na.rm = TRUE), t1=max(toa, na.rm = TRUE),
                        vertex.dyn.attrs = NULL, vertex.static.attrs= NULL,
+                       id.and.per.vars = NULL,
                        graph.attrs = NULL,
                        undirected=getOption("diffnet.undirected"),
                        self=getOption("diffnet.self"), multiple=getOption("diffnet.multiple")) {
 
-  # Step 0.0: Check if its diffnet!
+  # Step 0.0: Check if its diffnet! --------------------------------------------
   if (inherits(graph, "diffnet")) {
     message("Nothing to do, the graph is already of class \"diffnet\".")
     return(graph)
@@ -179,57 +540,7 @@ as_diffnet <- function(graph, toa, t0=min(toa, na.rm = TRUE), t1=max(toa, na.rm 
                                 meta$n, " and ", length(toa), " respectively). ",
                                 "-toa- should be of length n (number of vertices).")
 
-  # Vertex attrs
-  if (length(vertex.dyn.attrs)) {
-
-    attlen <- lapply(vertex.dyn.attrs, nrow)
-    if (any(attlen != meta$n)) stop("-graph- and -vertex.dyn.attrs- have different lengths (",
-                                    meta$n, " and ", paste(attlen, collapse=", "), "respectively). ",
-                                    "-vertex.dyn.attrs- should have n rows.")
-
-    # Coercing into data.frames
-    cnames <- colnames(vertex.dyn.attrs[[1]])
-    if (!length(cnames))
-      cnames <- sprintf("v.dyn.att%03d", 1:ncol(vertex.dyn.attrs[[1]]))
-
-    # Checking if it is a data.frame or not
-    if (!inherits(vertex.dyn.attrs[[1]], "data.frame"))
-      warning("-vertex.dyn.attrs- will be coerced to a data.frame.")
-
-    vertex.dyn.attrs <- lapply(vertex.dyn.attrs, function(x) {
-      if (!inherits(x, "data.frame")) x <- as.data.frame(x)
-      dimnames(x) <- list(meta$ids, cnames)
-      x
-    })
-  }
-
-  # Checking static attributes
-  if (length(vertex.static.attrs)) {
-
-    # Must be either a data.frame or a matrix
-    if (!any(c("data.frame", "matrix") %in% class(vertex.static.attrs)))
-      stop("-vertex.static.attrs- must be either a data.frame or a matrix.")
-
-    attlen <- nrow(vertex.static.attrs)
-    if (attlen != meta$n) stop("-graph- and -vertex.static.attrs- have different lengths (",
-                               meta$n, " and ", attlen, "respectively). ",
-                               "-vertex.static.attrs- should have n rows.")
-
-    # Coercing into matrix
-    cnames <- colnames(vertex.static.attrs)
-    if (!length(cnames))
-      cnames <- sprintf("v.static.att%03d", 1:ncol(vertex.static.attrs))
-
-    # Checking if it is a data.frame or not
-    if (!inherits(vertex.static.attrs, "data.frame")) {
-      warning("-vertex.static.attrs- will be coerced to a data.frame.")
-      vertex.static.attrs <- as.data.frame(vertex.static.attrs)
-    }
-
-    dimnames(vertex.static.attrs) <- list(meta$ids, cnames)
-  }
-
-  # Step 2.1: Checking class of TOA and coercing if necesary
+  # Step 2.1: Checking class of TOA and coercing if necesary -------------------
   if (!inherits(toa, "integer")) {
     warning("Coercing -toa- into integer.")
     toa <- as.integer(toa)
@@ -239,7 +550,7 @@ as_diffnet <- function(graph, toa, t0=min(toa, na.rm = TRUE), t1=max(toa, na.rm 
   if (!length(names(toa)))
     names(toa) <- meta$ids
 
-  # Step 3.1: Creating Time of adoption matrix -----------------------------------
+  # Step 3.1: Creating Time of adoption matrix ---------------------------------
   mat <- toa_mat(toa, labels = meta$ids, t0=t0, t1=t1)
 
   # Step 3.2: Verifying dimensions and fixing meta$pers
@@ -256,7 +567,22 @@ as_diffnet <- function(graph, toa, t0=min(toa, na.rm = TRUE), t1=max(toa, na.rm 
 
   meta$pers <- as.integer(colnames(mat$adopt))
 
-  # Step 4.1: Change the class (or set the names) of the graph
+  # Step 4.0: Checking the attributes ------------------------------------------
+
+  # Vertex dyn attrs
+  vertex.dyn.attrs <-
+    check_as_diffnet_attrs(vertex.dyn.attrs, meta, TRUE, id.and.per.vars)
+
+  # Vertex static attrs
+  vertex.static.attrs <-
+    check_as_diffnet_attrs(vertex.static.attrs, meta, FALSE, id.and.per.vars)
+
+  # Vertex static attrs
+  graph.attrs <-
+    check_as_diffnet_attrs(graph.attrs, meta, FALSE, id.and.per.vars, "graph")
+
+
+  # Step 4.1: Change the class (or set the names) of the graph -----------------
   if (meta$class=="array") {
     graph <- lapply(1:meta$nper, function(x) {
       x <- methods::as(graph[,,x], "dgCMatrix")
@@ -275,25 +601,6 @@ as_diffnet <- function(graph, toa, t0=min(toa, na.rm = TRUE), t1=max(toa, na.rm 
   meta$self       <- self
   meta$undirected <- undirected
   meta$multiple   <- multiple
-
-  # Checking attributes (last step)
-  if (!length(vertex.dyn.attrs)) {
-    vertex.dyn.attrs <-
-      lapply(meta$pers, function(x) {
-        as.data.frame(
-          matrix(ncol=0, nrow=meta$n, dimnames = list(meta$ids, NULL))
-          )
-      })
-
-    # Labeling
-    names(vertex.dyn.attrs) <- meta$pers
-  }
-
-  if (!length(vertex.static.attrs)) {
-    vertex.static.attrs <- as.data.frame(
-      matrix(ncol=0, nrow=meta$n, dimnames = list(meta$ids, NULL))
-    )
-  }
 
   return(structure(list(
     graph = graph,
@@ -341,7 +648,7 @@ diffnet.attrs <- function(graph, element=c("vertex","graph"), attr.class=c("dyn"
   if (!length(v.static)) v.static <- as.data.frame(matrix(ncol=0, nrow=n))
   if (!length(v.dyn[[1]])) v.dyn <- lapply(1:nper, function(y) as.data.frame(matrix(ncol=0, nrow=n)))
 
-  attrs <- cbind(toa=graph$toa, v.static, g.static)
+  attrs <- cbind(toa=graph$toa, v.static)
   out <- lapply(1:nper, function(y) {
     cbind(per=rep(pers[y], n),attrs, v.dyn[[y]])
   })