+++ R/stats.R
@@ -191,7 +191,9 @@ dgr.array <- function(graph, cmode, undirected, self, valued) {
 #' the element-wise inverse of the structural equivalence matrix (see example below).
 #' Furthermore, if \code{alt.graph="se"}, the inverse of the structural equivalence
 #' is computed via \code{\link{struct_equiv}} and used instead of the provided
-#' graph.
+#' graph. Notice that when using a valued graph the option \code{valued} should
+#' be equal to \code{TRUE}, this check is run automatically when running the
+#' model using structural equivalence.
 #'
 #' If the user does not specifies a particular weighting attribute in \code{attrs},
 #' the function sets this as a matrix of ones. Otherwise the function will return
@@ -223,7 +225,9 @@ dgr.array <- function(graph, cmode, undirected, self, valued) {
 #'
 #' SE <- lapply(struct_equiv(graph), "[[", "SE")
 #' SE <- lapply(SE, function(x) 1/(x + 1e-15))
-#' expo_se <- exposure(graph, alt.graph=SE)
+#'
+#' # Recall setting valued equal to TRUE!
+#' expo_se <- exposure(graph, alt.graph=SE , valued=TRUE)
 #'
 #' # These three lines are equivalent to
 #' expo_se2 <- exposure(graph, alt.graph="se")
@@ -244,7 +248,7 @@ dgr.array <- function(graph, cmode, undirected, self, valued) {
 #' @export
 #' @author Vega Yon, Dyal, Hayes & Valente
 exposure <- function(graph, cumadopt, attrs = NULL, alt.graph=NULL, outgoing=TRUE,
-                     valued=getOption("diffnet.valued"), normalized=TRUE, ...) {
+                     valued=getOption("diffnet.valued", FALSE), normalized=TRUE, ...) {
 
   # Checking diffnet attributes
   if (length(attrs) == 1 && inherits(attrs, "character")) {
@@ -281,7 +285,15 @@ exposure <- function(graph, cumadopt, attrs = NULL, alt.graph=NULL, outgoing=TRU
       # Computing structural equivalence
       se <- lapply(struct_equiv(graph,...), "[[", "SE")
       se <- lapply(se, function(x) 1/(x + 1e-15))
+
+      # Changing valued
+      if (!valued) {
+        warning("To use alt.graph=\"se\" -valued- has been switched to TRUE.")
+        valued <- TRUE
+      }
+
       se
+
     } else alt.graph
   }